+++ changelog.md
@@ -3,6 +3,7 @@
 + Rename "Coordinate" to "Coord"(keep old name).
 + Add "sx" & "sy" properties for VPoint slot coordinate.
 + Add "link_pos" and "points_pos" methods for VPoint and VLink respectively.
++ Add "to_coord" method for VPoint objects.
 
 # Version 20.06